html/semantics/embedded-content-0/media-elements/error_src_not_supported.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.error - MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-error">spec reference</a></p>
  <audio id="a" src=".">
  </audio>
  <video id="v" src=".">
  </video>
  <div id="log"></div>
  <script>
var tv = async_test("video.error - MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source");
var v = document.getElementById("v");
v.onerror = function() {
  tv.step(function() {
    assert_equals(
     v.error.code,
     v.error.MEDIA_ERR_SRC_NOT_SUPPORTED,
     "videoElement.error property is MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source");
  });
  tv.done();
}

var ta = async_test("audio.error - MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source");
var a = document.getElementById("a");
a.onerror = function() {
  ta.step(function() {
    assert_equals(
     a.error.code,
     a.error.MEDIA_ERR_SRC_NOT_SUPPORTED,
     "audioElement.error property is MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source");
  });
  ta.done();
}
  </script>
 </body>
</html>
